ability to add images to create new product
