+++ react-app/src/components/ListingForm/UploadPicture.js
@@ -1,47 +1,26 @@
-import React, { useState } from 'react';
+import React, { useState, useEffect } from 'react';
 import { useHistory } from 'react-router-dom';
+import productsReducer from '../../store/products';
 import './ListingForm.css';
 
-const UploadPicture = () => {
+const UploadPicture = ({ images, setImages }) => {
 	const history = useHistory(); // so that we can redirect after the image upload is successful
-	const [image, setImage] = useState(null);
 	const [imageLoading, setImageLoading] = useState(false);
-
-	const handleSubmit = async (e) => {
-		e.preventDefault();
-		const formData = new FormData();
-		formData.append('image', image);
-
-		// aws uploads can be a bit slow—displaying
-		// some sort of loading message is a good idea
-		setImageLoading(true);
-
-		const res = await fetch('/api/images', {
-			method: 'POST',
-			body: formData,
-		});
-		if (res.ok) {
-			await res.json();
-			setImageLoading(false);
-		} else {
-			setImageLoading(false);
-			// a real app would probably use more advanced
-			// error handling
-			console.log('error');
-		}
-	};
 
 	const updateImage = (e) => {
 		const file = e.target.files[0];
-		setImage(file);
+		setImages([...images, file]);
 	};
 
 	return (
-		<form onSubmit={handleSubmit}>
+		<>
 			<input type='file' accept='image/*' onChange={updateImage} />
-			<button type='submit'>Submit</button>
-			{imageLoading && <p>Loading...</p>}
-		</form>
+			<div id='picture-preview-container'>
+				{images.map((image, i) => (
+					<img key={i} src={URL.createObjectURL(image)} style={{ maxWidth: '300px' }} />
+				))}
+			</div>
+		</>
 	);
 };
