+++ react-app/src/components/ListingForm/index.js
@@ -1,5 +1,5 @@
 import './ListingForm.css';
-import { useState } from 'react';
+import { useState, useEffect } from 'react';
 import { useDispatch } from 'react-redux';
 import { createProduct } from '../../store/products';
 import UploadPicture from './UploadPicture';
@@ -21,8 +21,9 @@ export default function ListingForm({ product, userId, setShowForm }) {
 	const [quantity, setQuantity] = useState(product?.quantity || 1);
 	const [productType, setProductType] = useState(product?.product_type_id || 1);
 	const [petType, setPetType] = useState(product?.pet_type_id || 1);
+	const [images, setImages] = useState([]);
 
-	const handleSubmit = (e) => {
+	const handleSubmit = async (e) => {
 		e.preventDefault();
 
 		const detailsArr = [];
@@ -43,9 +44,32 @@ export default function ListingForm({ product, userId, setShowForm }) {
 			pet_type_id: +petType,
 		};
 
-		console.log(newProduct);
+		const createdProduct = await dispatch(createProduct(newProduct));
 
-		dispatch(createProduct(newProduct));
+		images.forEach(async (image) => {
+			const formData = new FormData();
+			formData.append('image', image);
+			formData.append('product_id', createdProduct.id);
+
+			// aws uploads can be a bit slow—displaying
+			// some sort of loading message is a good idea
+			// setImageLoading(true);
+
+			const res = await fetch('/api/images', {
+				method: 'POST',
+				body: formData,
+			});
+			if (res.ok) {
+				await res.json();
+				// setImageLoading(false);
+			} else {
+				// setImageLoading(false);
+				// a real app would probably use more advanced
+				// error handling
+				console.log('error');
+			}
+		});
+
 		setShowForm(false);
 	};
 
@@ -74,13 +98,11 @@ export default function ListingForm({ product, userId, setShowForm }) {
 		console.log(editedProduct);
 	};
 
-	const uploadImage = (e) => {
-		console.log(e);
-	};
-
 	return (
 		<>
-			<UploadPicture />
+			<div id='picture-preview'>
+				<UploadPicture images={images} setImages={setImages} />
+			</div>
 			<form id='listingForm' onSubmit={product ? handleEdit : handleSubmit}>
 				<label>
 					Title
